Update README.md
This is a python mini project - Book Library. This programe allows the user to see the book list and rent, sell or return a book in the library. The books that the library purchase are predefined along with the prices and rents of the book. There is also a membership feature that will allow you to rent two books at a time (while the normal customer can only rent one book at a time)and extend the time of renting. The code is able to figure out whether a person had rented a book previously and whether the VIP code entered belongs to the customer name or not.